bugfix: update UniqueLogicalSize from int64 to float64

+++ models/nvmf_namespace.go
@@ -169,7 +169,7 @@ type NvmfNamespace struct {
 	ThinProvision *bool `json:"thin_provision"`
 
 	// unique logical size
-	UniqueLogicalSize *int64 `json:"unique_logical_size,omitempty"`
+	UniqueLogicalSize *float64 `json:"unique_logical_size,omitempty"`
 
 	// unique size
 	// Required: true
